refactor: Smf -> SMF
use consistent file naming
