+++ src/gui/src/ExportMidiDialog.cpp
@@ -30,7 +30,7 @@
 #include <core/Basics/Song.h>
 #include <core/Hydrogen.h>
 #include <core/Preferences/Preferences.h>
-#include <core/Smf/SMF.h>
+#include <core/SMF/SMF.h>
 
 using namespace H2Core;
